requirements.txt
numpy
pandas
matplotlib
jupyter
pyserial
pyvisa
setuptools

 # requirements.txt
 #
 # installs dependencies from ./setup.py, and the package itself,
 # in editable mode
 -e .